pyproject.toml
[tool.poetry]
name = "api-text-tone"
version = "0.1.0"
description = ""
authors = ["misha.mironov <spaik.mironov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
streamlit = "^1.29.0"
transformers = "^4.35.2"
pytest = "^7.4.3"
pydantic = "^2.5.2"
torch = "^2.1.1"
jinja2 = "^3.1.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
